requirements.txt
# Core LangChain and LangGraph Orchestration
langchain-core
langchain-community
langgraph
langchain

# LLM Provider (You can swap this for others like anthropic, google-genai, etc.)
openai
langchain-openai
# API Framework and Server
fastapi
uvicorn

#Database
sqlalchemy
psycopg2-binary

# Environment Management
python-dotenv
pydantic

# Telegram Integration
python-telegram-bot[webhooks]
httpx # For making internal API calls and Google OAuth

# Authentication & Security
python-jose[cryptography]  # JWT token handling
passlib[bcrypt]            # Password hashing with bcrypt
redis                      # Redis caching for sessions and JWT blacklist